requirements.txt
GitPython==3.1.43
lxml==6.0.0
kometautils==0.3.4
requests==2.31.0